# perf: cache node resolution when accesing a global