--- a/cli/tests/testdata/run/with_package_json/npm_binary/main.out
+++ b/cli/tests/testdata/run/with_package_json/npm_binary/main.out
@@ -1,9 +1,6 @@
 [WILDCARD]package.json file found at '[WILDCARD]with_package_json[WILDCARD]npm_binary[WILDCARD]package.json'
 [WILDCARD]
 this
-[WILDCARD]
 is
-[WILDCARD]
 a
-[WILDCARD]
 test