Reclaims some of the performance hit introduced by https://github.com/denoland/deno/pull/19307.

Express

`v1.35.1`

```
Running 10s test @ http://localhost:7001
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   366.00us  280.04us  10.70ms   96.05%
    Req/Sec    14.33k     1.16k   15.79k    89.11%
  Latency Distribution
     50%  317.00us
     75%  345.00us
     90%  408.00us
     99%    1.50ms
  288027 requests in 10.10s, 58.78MB read
Requests/sec:  28513.16
Transfer/sec:      5.82MB
```

`v1.35.2`

```
Running 10s test @ http://localhost:7001
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   730.26us  267.97us  10.78ms   95.44%
    Req/Sec     6.91k   413.47     7.25k    90.59%
  Latency Distribution
     50%  686.00us
     75%  703.00us
     90%  771.00us
     99%    1.58ms
  138852 requests in 10.10s, 28.34MB read
Requests/sec:  13747.35
Transfer/sec:      2.81MB
```

this PR
```
Running 10s test @ http://localhost:7001
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   393.91us  268.28us   9.46ms   95.53%
    Req/Sec    13.23k     1.07k   14.64k    83.17%
  Latency Distribution
     50%  343.00us
     75%  372.00us
     90%  445.00us
     99%    1.51ms
  265928 requests in 10.10s, 54.27MB read
Requests/sec:  26326.57
Transfer/sec:      5.37MB
```

`npm:ws` 

`v1.35.1`

```
./third_party/prebuilt/mac/load_test 10 0.0.0.0 8080 0 0
Running benchmark now...
Msg/sec: 118798.250000
```

`v1.35.2`

```
./third_party/prebuilt/mac/load_test 10 0.0.0.0 8080 0 0
Running benchmark now...
Msg/sec: 11843.500000
Msg/sec: 12015.750000
```

this PR
```
./third_party/prebuilt/mac/load_test 10 0.0.0.0 8080 0 0
Running benchmark now...
Msg/sec: 95359.000000
Msg/sec: 97625.000000
```